Refactor `getIsPostTypeHierarchical` to utilities functions
